adding logistic activation function
