+++ union_pooling/union_pooling/activation_strategies/activation_function_base.py
@@ -1,6 +1,12 @@
 from abc import ABCMeta, abstractmethod
 
-class ActivationFunctionBase(metaclass=ABCMeta):
+class ActivationFunctionBase(object):
+
+  __metaclass__ = ABCMeta
+
+
+  def __init__(self):
+    pass
 
 
   @abstractmethod
